make ShrinkQueue not scaleset specific
This pulls forwards the ShrinkQueue changes from the autoscale PR
